requirements.txt
discord.py>=2.3.0
anthropic>=0.39.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
sqlalchemy>=2.0.0
aiomysql>=0.2.0
PyMySQL>=1.1.0

